added line to remove /opt/sandstorm to uninstall
On an Ubuntu 16.04-LTS server, uninstalling sandstorm using 'sandstorm uninstall --delete-user-data' & then reinstalling sandstorm throws up an error that says Sandstorm is still installed. It appears this error is triggered because the /opt/sandstorm directory hasn't been removed / deleted during the uninstall process. 

So I added a line that asks the users to verify that the /opt/sandstorm directory has been removed, and if it is still on the drive, use 

sudo rm -r /opt/sandstorm